+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 * feat(ChatMessage): add `files` slot
 * feat(EditorSuggestionMenu): expose suggestion matching options
 * feat(Select): support `item-aligned` position mode
+* feat(components): resolve `defaultVariants` in template logic
 * 
 * fix(Error): support `status` and `statusText` properties
 * fix(ContentSurround): handle RTL mode